Made the style a bit better.
There is no more overlapping with the player and the other tables. When
the screen is too small to show the whole gui, a section will appear
under the other sections. Everything should be based on pixels, instead
of percents, to provide a more uniform style.

The player doesn't scoll down on the page anymore, but I didn't think
that was too necessary.