Convert project to maven and adds the launcher project
This commit shuffles around source files and adds maven pom.xml files with
dependencies.

The ScreenCastingLocalHub project was moved to a subproject, and the launcher
was added as a new subproject. The top level maven pom.xml is the parent of the
two projects.

The dependency management provided my Maven makes the externalJars
directory obsolete and it can probably be removed, but it is left in case I
missed something. (Not every jar in externalJars was turned in to a maven
dependency because not all appear to be used)

Note that a local maven repository was added in the ScreenCastingLocalHub/repo
directory to hold the sslcerts.jar dependency.

This commit also copies in the ScreenRecordingModule.java file and its
supporting files from the other repository.

Also note that there was an old "launcher" in the launcher directory. It has
been removed.

And finally, this adds a travis config for continuous integration support.